requirements.txt
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-HTTPAuth==4.2.0
streamlink==2.2.0
youtube-dl==2021.6.6
urllib3==1.23